Moved courses/gradebook to courses/gradebook/fetch
