+++ internal/api/courses/routes.go
@@ -6,6 +6,7 @@ import (
 	"github.com/edulinq/autograder/internal/api/core"
 	"github.com/edulinq/autograder/internal/api/courses/admin"
 	"github.com/edulinq/autograder/internal/api/courses/assignments"
+	"github.com/edulinq/autograder/internal/api/courses/gradebook"
 	"github.com/edulinq/autograder/internal/api/courses/lms"
 	"github.com/edulinq/autograder/internal/api/courses/stats"
 	"github.com/edulinq/autograder/internal/api/courses/upsert"
@@ -14,7 +15,6 @@ import (
 
 var baseRoutes []core.Route = []core.Route{
 	core.MustNewAPIRoute(`courses/get`, HandleGet),
-	core.MustNewAPIRoute(`courses/gradebook`, HandleGradebook),
 	core.MustNewAPIRoute(`courses/list`, HandleList),
 }
 
@@ -24,6 +24,7 @@ func GetRoutes() *[]core.Route {
 	routes = append(routes, baseRoutes...)
 	routes = append(routes, *(admin.GetRoutes())...)
 	routes = append(routes, *(assignments.GetRoutes())...)
+	routes = append(routes, *(gradebook.GetRoutes())...)
 	routes = append(routes, *(lms.GetRoutes())...)
 	routes = append(routes, *(stats.GetRoutes())...)
 	routes = append(routes, *(upsert.GetRoutes())...)
